+++ exceptions/exceptions.py
